==> code/01_sign_in.R <==
# Work in progress. DRAFT.


library(twitteR)


source("~/Documents/Div/credentials/twitter_oauth.R")


setup_twitter_oauth(consumer_key = consumerKey,
                    consumer_secret = consumerSecret,
                    access_token = accessToken,
                    access_secret = accessSecret)
