==> code/04_prepare_tweets.Rmd <==
---
title: "Data Preparation"
author: "Sebastian Sauer"
date: "5 8 2017"
output: html_document
---




```{r setup, include=FALSE}

opts_knit$set(root.dir = normalizePath('../'))


knitr::opts_chunk$set(echo = TRUE,
                      message = FALSE,
                      fig.align = "center",
                      cache = TRUE,
                      root.dir = )
```


---

 Work in progress. DRAFT.

---



**Attention** Raw data is being touched. Watch out, do not mess.


#tl;dr

This script cleanses and prepares the tweet data for further analysis.
Input: 
`german_politicians_twitter_raw.csv` -- Twitter accounts
`tweets_polits_raw_with_trump.Rdata`  -- Tweets

Output:
`polits_df.Rdata` -- Twitter Accounts
`tweets_df.Rdata` -- Tweets
`party_pal.Rdata` -- Color scheme


# Warmup


First load some libraries.

```{r}
library(tidyverse)
library(readr)
library(lubridate)
library(stringr)
library(testthat)
```

## Load data

And the data - raw data.

```{r load-raw-data}
getwd()

polits_df <- read_csv("../data_polit_twitter/raw/german_politicians_twitter_raw.csv")

load("../data_polit_twitter/raw/tweets_polits_raw_with_trump.Rdata")
```




## Checks

```{r}
polits_df <- filter(polits_df, !is.na(screenName))

if (any(is.na(polits_df$screenName))) cat("There are screenNames missing!")
if (any(is.na(polits_df$party))) cat("There are party names missing!")

polits_df %>% 
  filter(str_detect(screenName,"hahn")) %>% 
  select(screenName, party)
```


## Check for Non-Ascii

Watch out! There are some hidden non-Ascii characters in Twitter accounts. For example a character which looks as "ff" - but it's ONE character, albeit not an ASCII one. Remove such characters before proceeding! Do that manually in the raw accounts file. Look out for:

- ff
- fl
- fi


```{r}
tools::showNonASCII(polits_df$screenName)
```

There should be NONE non-Ascii character. Only ASCII characters are valid.



## Remove non-human accounts

Accounts such as "gruene_bt" should be removed.

```{r}
polits_df %>% 
  filter(human == "y") -> polits_df
```



# Join parties-df 



## Convert to lower_case
```{r}

polits_df %>% 
  mutate(party = str_to_lower(party)) -> polits_df
```

Check if the screename is *not* preceded by an asterisk '@':

```{r}
polits_df %>% pull(screenName) %>% head
```

Otherwise, change in raw data file (exclude '@').


Glimpse

```{r}
polits_df %>% 
  glimpse
```

## Join

```{r}
polits_df %>% 
  dplyr::select(screenName, party) -> polits_parties

tweets_polits_raw_with_trump %>% 
  right_join(polits_parties, by = "screenName") -> tweets_polits

names(tweets_polits)

```


## Check


```{r}

if (any(is.na(tweets_polits$screenName))) {
  cat("There are screenNames missing!") } else {
    cat("There are NO screenNames missing. OK.")
  }

if (!any(is.na(tweets_polits$party))) {
  cat("There are party NO names missing. OK.")} else {
     cat("There are party names missing!\n")
    tweets_polits %>% 
      filter(is.na(party)) %>% 
      select(screenName, party) %>% 
      distinct()
  }


```


Count the number of different screenNames

```{r}
polits_df %>% 
  select(screenName) %>% 
  n_distinct -> dummy1
dummy1

tweets_polits %>% 
  select(screenName) %>% 
  n_distinct -> dummy2
dummy2

expect_true(dummy1 == dummy2)

```



## Ensure UTF8

```{r}
Encoding(tweets_polits$text) <- "UTF8"

```



Glimpse:

```{r}
tweets_polits %>% glimpse
```




## How many tweets (initially)?

```{r}
tweets_polits %>% 
  nrow
```

~1*e+6 (1 Mio)


## Save this updated data (joined with party names)

```{r}
save(tweets_polits, file = "../data_polit_twitter/raw/tweets_polits_with_party_names.RData")
```



## Erase duplicates


needs some time...

```{r cache = TRUE}
tweets_polits %>% 
  group_by(id) %>% 
  filter(row_number() == 1) %>% 
  ungroup -> tweets_df

tweets_df %>% 
  nrow  # ~410k
```


## How many words in total?

```{r}
tweets_df$text %>% 
  head
```


```{r}
tweets_df$text %>% str_count %>% sum(., is.na = TRUE)
```

Approx ~50 Mio words


Proportion of tweets left (non-duplicates): `r nrow(tweets_df)/nrow(tweets_polits)`.



## Check for The Donald

```{r}
tweets_df %>% 
  count(party)
```





# Tweet duration

Get oldest and newst tweet per screenName. The difference between is the `tweet period`.

```{r}
tweets_df %>% 
  group_by(screenName) %>% 
  summarise(first_tweet = min(created),
            recent_tweet = max(created)) %>% 
  mutate(tweet_period = (recent_tweet - first_tweet) / ddays(1)) -> polits_tweet_times

names(polits_tweet_times)

```


Merge this information to the dataset of the politicians `polits_df`.

```{r}

names(polits_df)
names(polits_tweet_times)

polits_df %>% 
 left_join(polits_tweet_times, by = "screenName") -> polits_df

names(polits_df)

```


## Check


```{r}
if (any(is.na(polits_df$screenName))) cat("There are screenNames missing!")
if (any(is.na(polits_df$party))) cat("There are party names missing!")
```


Merge this to the `tweets_df` dataframe.

```{r}
names(tweets_df)
tweets_df %>% 
  select(-c(first_tweet, recent_tweet, tweet_period)) %>%
  left_join(polits_tweet_times, by = "screenName") -> tweets_df

names(tweets_df)
```



# Join the number of tweets to the `polits_df`


Join the info on the number of tweets to the polits_df.

```{r}
tweets_df %>% 
  group_by(screenName) %>% 
  summarise(n = n()) %>% 
  left_join(polits_df, by = "screenName") -> polits_df

names(polits_df)  
```


## Check


```{r}
if (any(is.na(polits_df$screenName))) cat("There are screenNames missing!")
if (any(is.na(polits_df$party))) cat("There are party names missing!")

polits_df %>% 
  select(screenName) %>% 
  n_distinct
```




# Daily tweets

```{r}
glimpse(polits_df)

polits_df %>% 
  mutate(daily_tweets_n = n / tweet_period,
         yearly_tweets_n = daily_tweets_n * 365) -> polits_df
```




# Correct typo

```{r}
tweets_df %>% 
  mutate(party = recode(party, Grune = "Gruene")) -> tweets_df


```


# Convert `party` to factor (and `str_to_lower`)

```{r}
tweets_df %>% 
  mutate(party = str_to_lower(party),
         party = factor(party)) -> tweets_df

names(tweets_df)
```


# Prepare color scalse for partis

```{r}
tweets_df$party %>% levels

party_pal <- c(afd = "dodgerblue3", 
               cdu = "grey2",
               csu = "steelblue2",
               fdp = "yellow1",
               fraktionslos = "grey70",
               gruene = "forestgreen",
               linke = "violet",
               spd = "red",
               trump = "tan1")
```



# Save data

```{r}
save(polits_df, file = "../data_polit_twitter/polits_df.Rdata")
save(tweets_df, file = "../data_polit_twitter/tweets_df.Rdata")
save(party_pal, file = "data/party_pal.Rdata")
```
