==> code/05_analyses_03_populism_score.Rmd <==
---
title: "Computing a populism measure"
author: "Sebastian Sauer"
date: "3 8 2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = normalizePath('../'))


knitr::opts_chunk$set(echo = TRUE,
                      message = FALSE,
                      fig.align = "center",
                      cache = TRUE)

save_to_disk <- FALSE
```


---

 Work in progress. DRAFT.

---


Let's perform some  analyses of polit tweets.

# Setup


First load some libraries.

```{r}
library(tidyverse)
library(readr)
library(lubridate)
library(magrittr)
library(tidytext)
library(stringr)
library(viridis)
library(wordcloud)
library(SnowballC)
library(knitr)
```


And the data.
```{r load-data}
load("../data_polit_twitter/tweets_df.Rdata")
load("../data_polit_twitter/tweet_tokens.Rdata")
load("../data_polit_twitter/polits_df.Rdata")
load("data/party_pal.Rdata")
load("../data_polit_twitter/party_pop_scores.Rdata")

```


Run this if some columns are matrices:

```{r}

polits_df %>% 
  mutate_all(funs(c)) -> polits_df

```


# Aspects of the populism measure


- Relative number of exclamation marks '!'
- Relative number of '1 / question marks' '?'
- Positivity score (score of 'Pos - Neg')
- Emotionality score (socre of 'Pos + Neg')
- Ratio of adjectives/adverbs
- ~~Mean number of characters per sentence~~
- Mean number of characters per word
- Relative number of semicolons ';'
- Relative number of words in CAPITAL LETTERS



# Get quotation/exclamation marks per policitian


```{r}
head(tweets_df)

tweets_df %>% 
  select(party, screenName, text) %>% 
  group_by(screenName) %>% 
  summarise(exclamation_marks_abs = str_count(text, pattern = "!") %>% sum,
            exclamation_marks_rel = exclamation_marks_abs / n(),
            party = first(party),
            complement_question_mark_rel = str_count(text, pattern = "\\?") %>% 
              sum %>% `/`(., n())) %>% 
  ungroup -> populism_scores
  
```


## Join with polits df 

First check, if not already joint

```{r}
populism_scores %>% 
  select(which(names(populism_scores) %in% names(tweets_df))) %>% 
  names

```



```{r eval = FALSE}
populism_scores %>% 
  select(-party)  %>%  # prevent duplicate column
  left_join(polits_df, by = "screenName") %>% 
  mutate(exclamation_prop = exclamation_marks_abs / word_count) -> polits_df
```




## Add relative number of semicolons

```{r}
tweets_df %>% 
  select(party, text, screenName) %>% 
  unnest_tokens(output = token, input = text) %>%  
  dplyr::filter(str_detect(token, ";")) %>% 
  select(-party) %>% 
  group_by(screenName) %>% 
  mutate(semicolon_n = n()) %>% 
  select(-token) %>% 
  full_join(polits_df) -> polits_df
```

I just leave the absolute number of semicolons, because only 1 account used semicolons at all.


# (Relative) number of words in CAPITAL LETTERS



```{r}
names(tweets_df)

tweets_df %>% 
  select(party, text, screenName) %>% 
  mutate(text = str_replace(text, "https*://\\S+", ""),  #remove links
         text = str_replace(text, "#\\S+", ""),  # remove hashtags
         rext = str_replace(text, "RT", ""),  # remove "RT"
         text = str_replace(text, "@\\S+", "")) %>%  # remove mentions
  #mutate(token = str_replace(text, " ", "")) %>% 
  #filter(str_detect(text, "[[:upper:]]{4,}")) %>% 
  mutate(cap_count = str_count(text, "[[:upper:]]{4,}")) %>% 
  dplyr::select(-party) %>% 
  group_by(screenName) %>% 
  summarise(cap_count = sum(cap_count, na.rm = T)) -> screenName_cap_count

select(polits_df, -cap_count) -> polits_df  

if ("cap_count" %in% names(polits_df)) {
  cat("column already present. Do not join again")
} else {
screenName_cap_count %>% 
full_join(polits_df, by = "screenName") %>%  
  mutate(cap_prop = cap_count / word_count) %>% 
  ungroup -> polits_df }


```





# Compute z-scores, and mean/md z-score


## Compute more indicators

```{r comp-indicators}
polits_df %>% 
  mutate(cap_prop = cap_count / word_count,
         neg_prop = neg_words_n / word_count,
         pos_prop = pos_words_n / word_count,
         adj_adv_quote = adjx_n / adv_n) -> polits_df


polits_df %>% ungroup -> polits_df

```


## Vector of pop scores

```{r}
pop_vars <- c("word_length_md_comp", 
              "emo_score_neg", 
              "emo_abs_score",
              "neg_words_ratio", 
              "cap_prop", 
              "neg_prop", 
              "adj_adv_quote",
              "emo_words_prop")


pop_vars_z <- sapply(pop_vars, function(col) paste0(col, "_z")) %>% as.character


length(pop_vars)

```



## Turn all indicators in right direction

Compute complements `comp` to change direction

```{r check-sign-of-pop-indicators}

polits_df %>% 
  mutate(semicolon_n = NULL,  # remove this variable, nearly all missing
         word_length_md_comp = 1 / word_length_md,
         emo_score_neg = -emo_score,  # flip signs
         emo_abs_score = emo_abs_score,
         neg_prop = neg_prop,
         cap_prop = cap_prop,
         adj_adv_quote = adj_adv_quote,
         emo_words_prop = neg_prop + pos_prop) %>% 
  ungroup -> polits_df
         
```



## Compute z scores of pop indicators

```{r}
polits_df %>% 
  mutate_at(pop_vars, 
            funs(z = as.vector(scale(.)))) -> polits_df
```



## Compute summary populism score


```{r eight-pop-indidcators}
polits_df %>% 
  rowwise() %>% 
  mutate(pop_z_md = median(c(word_length_md_comp_z, 
                             emo_score_neg_z, 
                             emo_abs_score_z,
                             neg_words_ratio_z, 
                             cap_prop_z, 
                             neg_prop_z, 
                             adj_adv_quote_z,
                             emo_words_prop_z), na.rm = T)) %>% 
  ungroup -> polits_df
```




```{r comp_party_pop_z}
polits_df %>% 
  group_by(party) %>% 
  summarise(pop_z_md_party = median(pop_z_md, na.rm = TRUE)) %>% 
  left_join(polits_df, by = "party") -> polits_df
```



# No more data manipulation below this line!

(In this file)

Data manipulation is now finished. `polits_df` is not changed by subsequent analysis of populisms computation


# Save data

```{r}
# save(populism_scores, file = "../data_polit_twitter/populism_scores.Rdata")
save(polits_df, file = "../data_polit_twitter/polits_df.Rdata")
save(party_pop_scores, file = "../data_polit_twitter/party_pop_scores.Rdata")
```


# Plot populismus


## Plot pop scores by party, total score


```{r p_pop_geom_col}
polits_df %>% 
  group_by(party) %>% 
  filter( party != "fraktionslos") %>% 
  summarise(pop_z_md = median(pop_z_md, na.rm = TRUE)) %>% 
  ggplot +
  aes(x = reorder(party, pop_z_md), y = pop_z_md, fill = party) +
  geom_col() +
  coord_flip() +
  labs(x = "Partei",
       y = "Populismuswert",
       fill = "Partei",
       caption = "Median über alle Populismus-Indikatoren") +
  scale_fill_manual(values = party_pal) -> p_party_pop_scores

p_party_pop_scores

if (save_to_disk == TRUE) ggsave(p_party_pop_scores, file = "img/party_pop_scores.pdf")

```



```{r p_pop_geom_boxplot}
polits_df %>% 
  group_by(party) %>% 
  filter( party != "fraktionslos") %>% 
  ggplot +
  aes(x = reorder(party, pop_z_md_party), y = pop_z_md, fill = party) +
  geom_boxplot() +
  coord_flip() +
  labs(x = "Partei",
       y = "Populismuswert (z-Wert)",
       fill = "Partei",
       caption = "Median über alle Populismus-Indikatoren") +
  scale_fill_manual(values = party_pal) -> p_party_pop_scores_boxplot

p_party_pop_scores_boxplot

if (save_to_disk == TRUE) ggsave(p_party_pop_scores_boxplot, file = "img/p_party_pop_scores_boxplot.pdf")

```



```{r p_pop_geom_boxplot_wo_trump}
polits_df %>% 
  group_by(party) %>% 
  filter(!(party %in% c("fraktionslos", "trump"))) %>% 
  ggplot +
  aes(x = reorder(party, pop_z_md_party), y = pop_z_md, fill = party) +
  geom_boxplot() +
  coord_flip() +
  labs(x = "Partei",
       y = "Populismuswert (z-Wert)",
       fill = "Partei",
       caption = "Median über alle Populismus-Indikatoren") +
  scale_fill_manual(values = party_pal) -> p_pop_geom_boxplot_wo_trump

p_pop_geom_boxplot_wo_trump

if (save_to_disk == TRUE) ggsave(p_pop_geom_boxplot_wo_trump, file = "img/p_pop_geom_boxplot_wo_trump.pdf")

```




## Plot individual accounts for each party

```{r}
polits_df %>% 
   group_by(party) %>% 
  filter( party != "fraktionslos") %>% 
  summarise(pop_z_md = median(pop_z_md, na.rm = TRUE)) -> parties_pop_z_md
```


```{r p_pop_points}
polits_df %>% 
  group_by(party, screenName) %>% 
  filter( party != "fraktionslos") %>% 
  summarise(pop_z_md = median(pop_z_md, na.rm = TRUE)) %>% 
  ggplot +
  geom_point(data = parties_pop_z_md, 
             aes(x = reorder(party, pop_z_md), y = pop_z_md), 
             shape = 5, size = 5, color = "grey20") +
  # aes(x = reorder(party, pop_z_md), y = pop_z_md, color = party) +
  geom_jitter(width = .1, aes(x = party, y = pop_z_md, color = party)) +
  coord_flip() +
  labs(x = "Partei",
       y = "Populismuswert (z-Wert)",
       color = "Partei",
       caption = "dargestellt sind Mediane über alle Populismus-Indikatoren") +
  scale_color_manual(values = party_pal) -> p_party_pop_scores_individual

p_party_pop_scores_individual

if (save_to_disk == TRUE) ggsave(p_party_pop_scores_individual, file = "img/p_party_pop_scores_individual.pdf")
if (save_to_disk == TRUE) ggsave(p_party_pop_scores_individual, file = "img/p_party_pop_scores_individual.png")

```



## Compute pop scores by party, details

```{r comp-party-pop-scores}
polits_df %>% 
  group_by(party) %>% 
  filter(party != "fraktionslos") %>% 
  summarise_at(vars(word_length_md_comp_z, 
                      emo_score_neg_z, 
                      emo_abs_score_z,
                      neg_words_ratio_z, 
                      cap_prop_z, 
                      neg_prop_z, 
                      adj_adv_quote_z,
                      emo_words_prop_z), 
               funs(mean, median), na.rm = T) %>% 
  mutate_all(funs(c)) %>% 
  ungroup -> party_pop_scores
  
```


## Pop row summaries (mean/median)

```{r comp-pop-scores-row-summaries}

# compute row summaries (total pop score)

party_pop_scores %>%   
  rowwise() %>% 
  mutate(pop_score_md = median(c(word_length_md_comp_z_median, 
                             emo_score_neg_z_median, 
                             emo_abs_score_z_median,
                             neg_words_ratio_z_median, 
                             cap_prop_z_median, 
                             neg_prop_z_median, 
                             adj_adv_quote_z_median,
                             emo_words_prop_z_median), na.rm = T),
        pop_score_mean = mean(c(word_length_md_comp_z_median, 
                             emo_score_neg_z_median, 
                             emo_abs_score_z_median,
                             neg_words_ratio_z_median, 
                             cap_prop_z_median, 
                             neg_prop_z_median, 
                             adj_adv_quote_z_median,
                             emo_words_prop_z_median), na.rm = T)) %>% 
    ungroup -> party_pop_scores

party_pop_scores %>% glimpse

```


## Plot again, different score computation (totals)

```{r plot-party-pop-scores}
party_pop_scores %>% 
  ggplot(aes(x = reorder(party, pop_score_md), y = pop_score_md, fill = party)) +
  geom_col() +
  scale_fill_manual(values = party_pal) +
  labs(x = "Partei",
       fill = "Partei",
       y = "Populismuswert",
       caption = "dargestellt sind Mediane aggregiert über Tweets und Accounts") +
  coord_flip() -> p_party_pop_scores_2

p_party_pop_scores_2

if (save_to_disk == TRUE) ggsave(p_party_pop_scores_2, file = "img/p_party_pop_scores_2.pdf")
```

For comparision, same data, different computation

```{r}
p_party_pop_scores
```

## Pop scores without Trump (per party)

```{r plot-party-pop-scores-without-trump}
party_pop_scores %>% 
  filter(!party %in% c("trump")) %>% 
  ggplot(aes(x = reorder(party, pop_score_md), y = pop_score_md, fill = party)) +
  geom_col() +
  scale_fill_manual(values = party_pal) +
  labs(x = "Partei",
       fill = "Partei",
       y = "Populismuswert",
       caption = "dargestellt sind Mediane aggregiert über Tweets und Accounts") +
  coord_flip() -> p_party_pop_scores_wo_trump

p_party_pop_scores_wo_trump

if (save_to_disk == TRUE) ggsave(p_party_pop_scores_wo_trump, file = "img/p_party_pop_scores_wo_trump.pdf")
```


## String of pop vars in German

```{r}
pop_vars_de_median <- c(
                 word_length_md_comp_z_median = "Wortkürze",
                 neg_words_ratio_z_median = "Verhältnis negativer/positiver Wörter",
                 neg_prop_z_median = "Anteil negativer Wörter",
                 emo_words_prop_z_median =  "Anteil emotionaler Wörter",
                 emo_score_neg_z_median =  "Wert an negativer Emotion",
                 emo_abs_score_z_median = "Wert an Emotionalität",
                 cap_prop_z_median = "Anteil GROSSBUCHSTABEN",
                 adj_adv_quote_z_median =  "Verhältnis Adjektive zu Adverbien")

pop_vars_de <- c(word_length_md_comp_z = "Wortkürze",
                 neg_words_ratio_z = "Verhältnis negativer/positiver Wörter",
                 neg_prop_z = "Anteil negativer Wörter",
                 emo_words_prop_z =  "Anteil emotionaler Wörter",
                 emo_score_neg_z =  "Wert an negativer Emotion",
                 emo_abs_score_z = "Wert an Emotionalität",
                 cap_prop_z = "Anteil GROSSBUCHSTABEN",
                 adj_adv_quote_z =  "Verhältnis Adjektive zu Adverbien")


```


## Party names in full

```{r}
party_names <- c(
  afd = "AfD",
  cdu = "CDU",
  csu = "CSU",
  fdp = "FDP",
  gruene = "B90/Grüne",
  linke = "Die Linke",
  spd = "SPD"
)
```



## Plot populism indicators per party (profile of indicators)


```{r p_party_pop_scores_details}
party_pop_scores %>% 
  select(contains("median"), party) %>% 
  gather(key = Indikator, value = z_Wert, -party) %>% 
  ggplot +
  aes(x = Indikator, y = z_Wert, fill = party) +
  geom_col() +
  facet_wrap(~party) +
  scale_fill_manual(values = party_pal) + 
  labs(caption = 
         "dargestellt sind Mediane über alle Tweets aller Politiker") +
  scale_x_discrete(labels = pop_vars_de_median) +
  coord_flip() -> p_party_pop_scores_details

p_party_pop_scores_details

if (save_to_disk == TRUE) ggsave(p_party_pop_scores_details, 
       file = "img/p_party_pop_scores_details.pdf", 
       width = 7,
       height = 5)

```


```{r p_party_pop_scores_details_boxplot}
polits_df %>% 
  select(party, screenName,
         word_length_md_comp_z,
         emo_score_neg_z,
         emo_abs_score_z,
         neg_words_ratio_z,
         cap_prop_z,
         neg_prop_z,
         adj_adv_quote_z,
         emo_words_prop_z,
         ) %>% 
 gather(key = Indikator, value = z_Wert, -c(screenName, party))  %>% 
 ggplot +
  aes(x = Indikator, y = z_Wert, fill = party) +
  geom_boxplot() +
  facet_wrap(~party) +
  scale_fill_manual(values = party_pal) + 
  scale_x_discrete(labels = pop_vars_de) +
  labs(y = "z-Wert") +
  coord_flip() -> p_party_pop_scores_details_boxplot

p_party_pop_scores_details_boxplot

if (save_to_disk == TRUE) ggsave(p_party_pop_scores_details_boxplot, file = "img/p_party_pop_scores_details_boxplot.pdf")


p_party_pop_scores_details_boxplot +
  scale_y_continuous(limits = c(-2,2)) -> 
  p_party_pop_scores_details_boxplot_limited

if (save_to_disk == TRUE) ggsave(p_party_pop_scores_details_boxplot_limited, file = "img/p_party_pop_scores_details_boxplot_limited.pdf")

```



## Plot populism indicators per party without Trump


```{r p_party_pop_scores_details_boxplot}
polits_df %>% 
  filter(party != "trump") %>% 
  select(party, screenName,
         word_length_md_comp_z,
         emo_score_neg_z,
         emo_abs_score_z,
         neg_words_ratio_z,
         cap_prop_z,
         neg_prop_z,
         adj_adv_quote_z,
         emo_words_prop_z,
         ) %>% 
 gather(key = Indikator, value = z_Wert, -c(screenName, party))  %>% 
 ggplot +
  aes(x = Indikator, y = z_Wert, fill = party) +
  geom_boxplot() +
  facet_wrap(~party) +
  scale_fill_manual(values = party_pal) + 
  scale_x_discrete(labels = pop_vars_de) +
  coord_flip() -> p_party_pop_scores_details_boxplot_no_trump

p_party_pop_scores_details_boxplot_no_trump

if (save_to_disk == TRUE) ggsave(p_party_pop_scores_details_boxplot_no_trump, file = "img/p_party_pop_scores_details_boxplot_no_trump.pdf")

```



```{r p_party_pop_scores_details_no_trump_point_limited_grey}
polits_df %>% 
  filter(!party %in% c("trump", "fraktionslos")) %>% 
  select(party, screenName,
         word_length_md_comp_z,
         emo_score_neg_z,
         emo_abs_score_z,
         neg_words_ratio_z,
         cap_prop_z,
         neg_prop_z,
         adj_adv_quote_z,
         emo_words_prop_z,
         ) %>% 
 gather(key = Indikator, value = z_Wert, -c(screenName, party))  -> polits_df_long


party_pop_scores_md_long %>% 
  mutate(Indikator = str_replace(Indikator, "_median", "")) -> party_pop_scores_md_long2
 
polits_df_long %>% 
  ggplot +
  aes(x = Indikator, y = z_Wert,) +
  geom_point(color = "grey80", position = "jitter") +
  facet_wrap(~party, nrow = 1, labeller = labeller(party = party_names)) +
  scale_color_manual(values = party_pal) + 
  scale_x_discrete(labels = pop_vars_de) +
  scale_y_continuous(limits = c(-1.5,1.5), breaks = c(-1,0,1)) +
  theme(legend.position = "none") +
  coord_flip() +
  labs(caption = 
       "dargestellt sind Mediane über alle Tweets aller Politiker",
       y = "z-Wert",
       color = "Partei") +
  geom_point(data = party_pop_scores_md_long2, aes(color = party), size = 4) +
  geom_line(data = party_pop_scores_md_long2, aes(color = party), group = 1) -> p_party_pop_scores_details_no_trump_point_limited_grey

p_party_pop_scores_details_no_trump_point_limited_grey

if (save_to_disk == TRUE) ggsave("img/p_party_pop_scores_details_no_trump_point_limited_grey.pdf",
       width = 8,
       height =5)
```





```{r p_party_pop_scores_details_no_trump_limited}
party_pop_scores %>% 
  select(contains("median"), party) %>% 
  filter(party != "trump") %>% 
  gather(key = Indikator, value = z_Wert, -party) %>% 
  ggplot +
  aes(x = Indikator, y = z_Wert, fill = party) +
  geom_col() +
  facet_wrap(~party) +
  scale_fill_manual(values = party_pal) + 
  labs(caption = 
         "dargestellt sind Mediane über alle Tweets aller Politiker") +
  scale_x_discrete(labels = pop_vars_de_median) +
  coord_flip() -> p_party_pop_scores_details_no_trump

p_party_pop_scores_details_no_trump

p_party_pop_scores_details_no_trump +
  scale_y_continuous(limits = c(-2,2)) ->
  p_party_pop_scores_details_no_trump_limited

if (save_to_disk == TRUE) ggsave(p_party_pop_scores_details_no_trump, 
       file = "img/p_party_pop_scores_details_no_trump.pdf", 
       width = 7,
       height = 5)


if (save_to_disk == TRUE) ggsave(p_party_pop_scores_details_no_trump_limited, 
       file = "img/p_party_pop_scores_details_no_trump_limited.pdf", 
       width = 7,
       height = 5)

party_pop_scores %>% 
  select(contains("median"), party) %>% 
  filter(party != "trump") %>% 
  gather(key = Indikator, value = z_Wert, -party) -> party_pop_scores_md_long
 
party_pop_scores_md_long 
  ggplot +
  aes(x = Indikator, y = z_Wert, color = party) +
  geom_line(group = 1, color = "grey80") +
  geom_point(size = 4) +
  facet_wrap(~party, nrow = 1) +
  scale_color_manual(values = party_pal) + 
  labs(caption = 
         "dargestellt sind Mediane über alle Tweets aller Politiker",
       y = "z-Wert",
       color = "Partei") +
  scale_x_discrete(labels = pop_vars_de_median) +
  theme(legend.position = "bottom") +
  scale_y_continuous(breaks = c(-1, 0, 1)) +
  coord_flip() -> p_party_pop_scores_details_no_trump_point

p_party_pop_scores_details_no_trump_point

if (save_to_disk == TRUE) ggsave(file = "img/p_party_pop_scores_details_no_trump_point.pdf",
       width = 7,
       height = 4)

```



# Plot most populist politicians

```{r}
polits_df %>% 
  filter(!is.na(party)) %>% 
  top_n(10, pop_z_md) %>% 
  ggplot +
  aes(x = reorder(screenName, pop_z_md), 
      y = pop_z_md, color = party) %>% 
  geom_point(size = 7) +
  coord_flip() +
  scale_color_manual(values = party_pal) +
  labs(x = "Name",
       y = "Populismuswert (z-Wert)",
       color = "Partei") -> p_pop_screenName


p_pop_screenName


if (save_to_disk == TRUE) ggsave(p_pop_screenName,
       file = "img/p_pop_screenName.pdf")

```

### Closer look to Karl_Lauterbach


```{r}
polits_df %>% 
  filter(screenName == "Karl_Lauterbach") %>% 
  select(pop_vars_z) %>% 
  gather(Indikator, Wert) %>% 
  ggplot +
  aes(x = reorder(Indikator, Wert), y = Wert) +
  geom_point(size = 5) +
  coord_flip() +
  scale_x_discrete(labels = pop_vars_de) +
  labs(x = "Indikator", y = "Z-Wert") -> p_most_pop_screenName


p_most_pop_screenName

if (save_to_disk == TRUE) ggsave(p_most_pop_screenName,
       file = "img/p_most_pop_screenName.pdf")
```


## Some tweets

```{r}
tweets_df %>% 
  filter(screenName == "Karl_Lauterbach") %>% 
  sample_n(10) %>% 
  pull(text) -> top_1_pop_account_sample_tweets

save(top_1_pop_account_sample_tweets, 
     file = "../data_polit_twitter/top_1_pop_account_sample_tweets.Rdata")
```
